+++ CHANGELOG.md
@@ -10,6 +10,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com)
 
 <!-- Keep this up to date! After a release, change the tag name to the latest release -->
 
+### Added
+
+- #2929 - Reports - added a new Report column type for Predicted Tags (Assets)
+
 ## 5.6.0 - 2023-02-02
 
 ### Added